+++ src/main/java/ch/ethz/matsim/discrete_mode_choice/model/nested/DefaultNest.java
@@ -1,13 +1,8 @@
 package ch.ethz.matsim.discrete_mode_choice.model.nested;
 
-import java.util.Collection;
-import java.util.LinkedList;
-
 public class DefaultNest implements Nest {
 	private final String name;
-	private final double scaleParameter;
-
-	private final Collection<Nest> children = new LinkedList<>();
+	private double scaleParameter;
 
 	public DefaultNest(String name, double scaleParameter) {
 		this.name = name;
@@ -19,16 +14,11 @@ public class DefaultNest implements Nest {
 	}
 
 	@Override
-	public Collection<Nest> getChildren() {
-		return children;
-	}
-
-	public void addChild(DefaultNest child) {
-		children.add(child);
-	}
-
-	@Override
 	public double getScaleParameter() {
 		return scaleParameter;
+	}
+
+	public void setScaleParameter(double scaleParameter) {
+		this.scaleParameter = scaleParameter;
 	}
 }