+++ src/main/java/ch/ethz/matsim/discrete_mode_choice/model/nested/DefaultNestedTripCandidate.java
@@ -1,0 +1,17 @@
+package ch.ethz.matsim.discrete_mode_choice.model.nested;
+
+import ch.ethz.matsim.discrete_mode_choice.model.trip_based.candidates.DefaultTripCandidate;
+
+public class DefaultNestedTripCandidate extends DefaultTripCandidate implements NestedUtilityCandidate {
+	private final Nest nest;
+
+	public DefaultNestedTripCandidate(double utility, String mode, Nest nest) {
+		super(utility, mode);
+		this.nest = nest;
+	}
+
+	@Override
+	public Nest getNest() {
+		return nest;
+	}
+}