+++ src/main/java/ch/ethz/matsim/discrete_mode_choice/model/nested/NestedLogitSelector.java
@@ -5,74 +5,71 @@ import java.util.LinkedList;
 import java.util.List;
 import java.util.Optional;
 import java.util.Random;
-import java.util.stream.Collectors;
 
 import ch.ethz.matsim.discrete_mode_choice.model.utilities.UtilityCandidate;
 import ch.ethz.matsim.discrete_mode_choice.model.utilities.UtilitySelector;
+import ch.ethz.matsim.discrete_mode_choice.model.utilities.UtilitySelectorFactory;
 
-public class NestedLogitSelector<T extends UtilityCandidate> implements UtilitySelector<T> {
-	private final List<T> candidates = new LinkedList<>();
-	private final List<NestedUtilityCandidate> nestedCandidates = new LinkedList<>();
+public class NestedLogitSelector implements UtilitySelector {
+	private final List<NestedUtilityCandidate> candidates = new LinkedList<>();
+	private final List<UtilityCandidate> plainCandidates = new LinkedList<>();
 
 	private final NestCalculator calculator;
 	private final Nest rootNest;
 
-	public NestedLogitSelector(Nest rootNest) {
-		calculator = new NestCalculator(rootNest, nests);
+	private final double minimumUtility;
+	private final double maximumUtility;
+
+	public NestedLogitSelector(NestStructure structure, double minimumUtility, double maximumUtility) {
+		this.minimumUtility = minimumUtility;
+		this.maximumUtility = maximumUtility;
+
+		this.calculator = new NestCalculator(structure);
+		this.rootNest = structure.getRoot();
 	}
 
 	@Override
-	public void addCandidate(T candidate) {
-		candidates.add(candidate);
+	public void addCandidate(UtilityCandidate candidate) {
+		NestedUtilityCandidate nestedCandidate;
+
+		if (candidate instanceof NestedUtilityCandidate) {
+			nestedCandidate = (NestedUtilityCandidate) candidate;
+		} else {
+			nestedCandidate = new WrappingNestedUtilityCandidate(candidate, rootNest);
+		}
+
+		// TODO: Add more verbosity here
+		if (nestedCandidate.getUtility() > minimumUtility) {
+			if (nestedCandidate.getUtility() < maximumUtility) {
+				candidates.add(nestedCandidate);
+				calculator.addCandidate(nestedCandidate);
+				plainCandidates.add(candidate);
+			}
+		}
 	}
 
 	@Override
-	public Optional<T> select(Random random) {
+	public Optional<UtilityCandidate> select(Random random) {
 		// I) If not candidates are available, give back nothing
 		if (candidates.size() == 0) {
 			return Optional.empty();
 		}
 
-		// II) Filter candidates that have a very low utility
-		List<T> filteredCandidates = candidates;
+		// II) Create a probability distribution over candidates
+		List<Double> density = new ArrayList<>(candidates.size());
 
-		if (considerMinimumUtility) {
-			filteredCandidates = candidates.stream() //
-					.filter(c -> c.getUtility() > minimumUtility) //
-					.collect(Collectors.toList());
-
-			if (filteredCandidates.size() == 0) {
-				logger.warn(String.format(
-						"Encountered choice where all utilities were smaller than %f (minimum configured utility)",
-						minimumUtility));
-				return Optional.empty();
-			}
-		}
-
-		// III) Build a nest calculator
-		List<NestedUtilityCandidate> nestedCandidates = new LinkedList<>();
-
-		for (T candidate : filteredCandidates) {
-			if (!(candidate instanceof NestedUtilityCandidate)) {
-				nestedCandidates.add((NestedUtilityCandidate) candidate);
-				calculator.addCandidate((NestedUtilityCandidate) candidate);
-			} else {
-				WrappingNestedUtilityCandidate nestedCandidate = new WrappingNestedUtilityCandidate(candidate,
-						rootNest);
-				nestedCandidates.add(nestedCandidate);
-				calculator.addCandidate(nestedCandidate);
-			}
-		}
-
-		// IV) Create a probability distribution over candidates
-		List<Double> density = new ArrayList<>(filteredCandidates.size());
-
-		for (NestedUtilityCandidate candidate : nestedCandidates) {
+		for (NestedUtilityCandidate candidate : candidates) {
 			double probability = calculator.calculateProbability(candidate);
+
+			if (!Double.isFinite(probability)) {
+				probability = 0.0; // Revise this, maybe it is better to cosntruct the process so we don't have
+									// them at all. TODO.
+			}
+
 			density.add(probability);
 		}
 
-		// IV) Build a cumulative density of the distribution
+		// III) Build a cumulative density of the distribution
 		List<Double> cumulativeDensity = new ArrayList<>(density.size());
 		double totalDensity = 0.0;
 
@@ -85,6 +82,23 @@ public class NestedLogitSelector<T extends UtilityCandidate> implements UtilityS
 		double pointer = random.nextDouble() * totalDensity;
 
 		int selection = (int) cumulativeDensity.stream().filter(f -> f < pointer).count();
-		return Optional.of(filteredCandidates.get(selection));
+		return Optional.of(plainCandidates.get(selection));
+	}
+
+	static public class Factory implements UtilitySelectorFactory {
+		private final NestStructure structure;
+		private final double minimumUtility;
+		private final double maximumUtility;
+
+		public Factory(NestStructure structure, double minimumUtility, double maximumUtility) {
+			this.structure = structure;
+			this.minimumUtility = minimumUtility;
+			this.maximumUtility = maximumUtility;
+		}
+
+		@Override
+		public UtilitySelector createUtilitySelector() {
+			return new NestedLogitSelector(structure, minimumUtility, maximumUtility);
+		}
 	}
 }