+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 # CHANGELOG
 
+- Add Nested Logit infrastructure
 - BC: Remove over-complicated generics for UtilitySelector/Factory
 - Update MATSim version, fix MainModeIdentifier and remove StageActivityTypes
 - Make initial plan elements available to mode choice process